Collections in API data objects are now initialized as empty by default to prevent NPEs

+++ src/main/java/net/boreeas/riotapi/rest/Player.java
@@ -28,7 +28,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Player {
     private int championId;
-    private int summonerId;
+    private long summonerId;
     private int teamId;
 
     public Team getTeam() {
